Ported the STOMP server implementation to use a "stomplet" API.
The stomplet API allows for a single channel handler to be installed in Netty and it abstracts away
all the Netty specific interaction. With a single handler, it is now much easier to understand the
order of execution and to make sure the pipeline is properly terminated. The same work needs
to be done on the client side.